requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.0
tqdm>=4.65.0
tiktoken>=0.5.0
